+++ walletkit/CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.11.0](https://github.com/worldcoin/walletkit/compare/v0.10.0...v0.11.0) - 2026-03-19
+
+### Added
+
+- [**breaking**] version bump to 0.6 with new RP signature ([#308](https://github.com/worldcoin/walletkit/pull/308))
+- export credential store as raw bytes ([#307](https://github.com/worldcoin/walletkit/pull/307))
+- expose TFH recovery agent address in defaults ([#299](https://github.com/worldcoin/walletkit/pull/299))
+- production defaults ([#305](https://github.com/worldcoin/walletkit/pull/305))
+
+### Fixed
+
+- *(walletkit-core)* fix flaky test_session_cache_ttl TTL off-by-one ([#303](https://github.com/worldcoin/walletkit/pull/303))
+
 ## [0.10.0](https://github.com/worldcoin/walletkit/compare/v0.9.1...v0.10.0) - 2026-03-12
 
 ### Added